(refactor) move json module
