+++ src/helpers/helper_log.rs
@@ -2,11 +2,11 @@ use crate::context::Context;
 #[cfg(not(feature = "no_logging"))]
 use crate::error::RenderError;
 use crate::helpers::{HelperDef, HelperResult};
+#[cfg(not(feature = "no_logging"))]
+use crate::json::value::JsonRender;
 use crate::output::Output;
 use crate::registry::Registry;
 use crate::render::{Helper, RenderContext};
-#[cfg(not(feature = "no_logging"))]
-use crate::value::JsonRender;
 #[cfg(not(feature = "no_logging"))]
 use log::Level;
 #[cfg(not(feature = "no_logging"))]
